Extract EventParams
